+++ database/migrations/2026_02_26_182837_add_batch_uuid_column_to_activity_log_table.php
@@ -8,13 +8,19 @@ class AddBatchUuidColumnToActivityLogTable extends Migration
 {
     public function up()
     {
-        Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
-            $table->uuid('batch_uuid')->nullable()->after('properties');
-        });
+        if (! Schema::hasColumn(config('activitylog.table_name'), 'batch_uuid')) {
+            Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->uuid('batch_uuid')->nullable()->after('properties');
+            });
+        }
     }
 
     public function down()
     {
+        if (! Schema::hasColumn(config('activitylog.table_name'), 'batch_uuid')) {
+            return;
+        }
+
         Schema::connection(config('activitylog.database_connection'))->table(config('activitylog.table_name'), function (Blueprint $table) {
             $table->dropColumn('batch_uuid');
         });
